replace units with equivalent NoArg
We should not need to provide any argument in case of NoArg but for that we need Overlapping families FWIW. Though, that means that we will be end up in using closed type families which in turn means HasGenerator is not extensible. But that is a trade-off to get sane usability.